working on env optimization
